decoder: Really stop decoding on opStop opcode
There was a break already on a `case opStop:` but break inside switch
only breaks out of this switch, not outer construct - looks like it was
a thinko.

So on opStop let's explicitly break out of instruction processing loop.

If we do not do so on e.g "I5\n.I7\n.N." input first call to .Decode()
returns None (from N.), not 5 and second call to Decode() panics (for
this see next patch).