Drop support for Go 1.9, Go 1.10 and Go 1.11
Under those go versions `go test` does not support `-benchtime Nx`
syntax, that we are going to use in the next patch, and fails with

    invalid value "1x" for flag -test.benchtime: time: unknown unit x in duration 1x

Maybe it is not a big deal and if it were just `-benchtime Nx` we could
try to invent and do something to try to support those old Go versions.
However I have upcoming patches that will require Go 1.18 for generics,
and from that point of view trying to keep support for Go version older
than Go 1.18 would be not practical.