Drop support for Go < 1.18
In the next patch we will need to use generics, which are only available
starting from Go 1.18 . That Go 1.18 was released in spring of 2022 -
almost 2.5 years ago - and is already EOL now. The older releases became
EOL even earlier compared to Go 1.18 and are generally no longer
supported by Go upstream. Dropping support for Go < 1.18 remains support
for Go 1.18 - 1.23, i.e. 6 releases 3 of which are already in end of
life state. So dropping support for Go < 1.18 should be, hopefully, ok.