*: Linkify documentation
So that whenever documentation references a type or method, it is
displayed as clickable link in godoc. In my view it is especially useful
to have reference from top-level package overview to things it
references and explains.

Here is, for example, how top-level documentation looks before this patch:

https://lab.nexedi.com/kirr/misc/-/raw/d6c5394c/ogórek/godoc-nolinks.png

and after this patch:

https://lab.nexedi.com/kirr/misc/-/raw/d6c5394c/ogórek/godoc-links.png

The markup syntax for how to turn text into links is explained at

https://tip.golang.org/doc/comment#links